+++ tests/integration/test_config_handling.py
@@ -15,7 +15,7 @@ from delayed_assert import expect, assert_expectations
 from six.moves import mock
 
 from examples.test_rp_logging import LOG_MESSAGE
-from tests import REPORT_PORTAL_SERVICE
+from tests import REPORT_PORTAL_SERVICE, REPORT_PORTAL_PACKAGE
 from tests.helpers import utils
 
 TEST_LAUNCH_ID = 'test_launch_id'
@@ -144,4 +144,100 @@ def test_rp_log_batch_payload_size(mock_client_init):
 
     constructor_args = mock_client_init.call_args_list[0][1]
     expect(constructor_args['log_batch_payload_size'] == log_size)
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+@mock.patch(REPORT_PORTAL_PACKAGE + '.config.warnings.warn')
+def test_rp_api_key(mock_warnings, mock_client_init):
+    api_key = 'rp_api_key'
+    variables = utils.DEFAULT_VARIABLES.copy()
+    variables.update({'rp_api_key': api_key}.items())
+
+    result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                    variables=variables)
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+
+    expect(mock_client_init.call_count == 1)
+
+    constructor_args = mock_client_init.call_args_list[0][1]
+    expect(constructor_args['api_key'] == api_key)
+    expect(mock_warnings.call_count == 0)
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+@mock.patch(REPORT_PORTAL_PACKAGE + '.config.warnings.warn')
+def test_rp_uuid(mock_warnings, mock_client_init):
+    api_key = 'rp_api_key'
+    variables = utils.DEFAULT_VARIABLES.copy()
+    del variables['rp_api_key']
+    variables.update({'rp_uuid': api_key}.items())
+
+    result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                    variables=variables)
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+
+    expect(mock_client_init.call_count == 1)
+
+    constructor_args = mock_client_init.call_args_list[0][1]
+    expect(constructor_args['api_key'] == api_key)
+    expect(mock_warnings.call_count == 1)
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+@mock.patch(REPORT_PORTAL_PACKAGE + '.config.warnings.warn')
+def test_rp_api_key_priority(mock_warnings, mock_client_init):
+    api_key = 'rp_api_key'
+    variables = utils.DEFAULT_VARIABLES.copy()
+    variables.update({'rp_api_key': api_key, 'rp_uuid': 'rp_uuid'}.items())
+
+    result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                    variables=variables)
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+
+    expect(mock_client_init.call_count == 1)
+
+    constructor_args = mock_client_init.call_args_list[0][1]
+    expect(constructor_args['api_key'] == api_key)
+    expect(mock_warnings.call_count == 0)
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+@mock.patch(REPORT_PORTAL_PACKAGE + '.config.warnings.warn')
+def test_rp_api_retries(mock_warnings, mock_client_init):
+    retries = 5
+    variables = utils.DEFAULT_VARIABLES.copy()
+    variables.update({'rp_api_retries': str(retries)}.items())
+
+    result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                    variables=variables)
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+
+    expect(mock_client_init.call_count == 1)
+
+    constructor_args = mock_client_init.call_args_list[0][1]
+    expect(constructor_args['retries'] == retries)
+    expect(mock_warnings.call_count == 0)
+    assert_expectations()
+
+
+@mock.patch(REPORT_PORTAL_SERVICE)
+@mock.patch(REPORT_PORTAL_PACKAGE + '.config.warnings.warn')
+def test_retries(mock_warnings, mock_client_init):
+    retries = 5
+    variables = utils.DEFAULT_VARIABLES.copy()
+    variables.update({'retries': str(retries)}.items())
+
+    result = utils.run_pytest_tests(['examples/test_rp_logging.py'],
+                                    variables=variables)
+    assert int(result) == 0, 'Exit code should be 0 (no errors)'
+
+    expect(mock_client_init.call_count == 1)
+
+    constructor_args = mock_client_init.call_args_list[0][1]
+    expect(constructor_args['retries'] == retries)
+    expect(mock_warnings.call_count == 1)
     assert_expectations()